+++ src/graph.py
@@ -65,7 +65,7 @@ def PrintStreetInfo(streets):
         print(edge[0], ":", edge[1], "->", edge[2], "=", edge[3])
     print()
 
-# Milestone 1 Show Case
+# Milestone 1 : Generate graph in matrix
 def Milestone1():
     maps = {}
     maps = LoadCoordinate("SthyrelestNode.txt")
@@ -74,13 +74,3 @@ def Milestone1():
     PrintStreetInfo(streets)
     graphMatrix = ConvertStreetsIntoGraph(streets, len(maps))
     PrintMatrix(graphMatrix, len(maps))
-
-# Driver for Graph
-# Milestone1()
-# maps = {}
-# maps = LoadCoordinate("SthyrelestNode.txt")
-# PrintCoordinateInfo(maps)
-# streets = LoadStreet("SthyrelestEdge.txt")
-# PrintStreetInfo(streets)
-# graphMatrix = ConvertStreetsIntoGraph(streets, len(maps))
-# PrintMatrix(graphMatrix, len(maps))
