+++ test/MatrixTestCase.txt
@@ -26,3 +26,9 @@
 [np.inf, np.inf, np.inf, np.inf, np.inf, np.inf, 6.324, 4.0, np.inf, 2.0, 3.162],
 [6.325, np.inf, 2.828, np.inf, np.inf, 2.828, np.inf, 4.472, 2.0, np.inf, 3.162],
 [np.inf, np.inf, np.inf, np.inf, np.inf, np.inf, np.inf, np.inf, 3.162, 3.162, np.inf]]
+
+[[np.inf, 4.472, np.inf, np.inf, np.inf],
+[4.472, np.inf, 8.0  , 6.324, np.inf],
+[np.inf, 8.0  , np.inf, 2.828, 6.324],
+[np.inf, 6.324, 2.828, np.inf, 4.0],
+[np.inf, np.inf, 6.324, 4.0  , np.inf]]
